+++ src/core/util/render_manager.cpp
@@ -17,7 +17,7 @@ double render_manager::paint_displays( const std::vector< display* >& displays )
 	// holding the lock is spent in paint().  So we measure this time as an estimate of
 	// how long per cycle we are holding the lock.
 	double start = time.elapsed();
-	for(int d=0; d<displays.size(); d++)
+	for(size_t d=0; d<displays.size(); d++)
 		displays[d]->paint();
 	double paint = time.elapsed() - start;
 
@@ -29,7 +29,7 @@ double render_manager::paint_displays( const std::vector< display* >& displays )
 		else if ( swap_thread_pool->size() < displays.size() )
 			swap_thread_pool->size_controller().resize( displays.size()-1 );
 		
-		for(int d=1; d<displays.size(); d++)
+		for(size_t d=1; d<displays.size(); d++)
 			swap_thread_pool->schedule( boost::bind( &display::swap, displays[d] ) );
 	}
 	displays[0]->swap();
